Bug: Yielding in a hook stops in the wrong instruction [e288c5a91883793d14ed9e9d93464f6ee0b08915]
Yielding in a hook must decrease the program counter, because it already
counted an instruction that, in the end, was not executed. However,
that decrement should be done only when about to restart the thread.
Otherwise, inspecting the thread with the debug library shows it one
instruction behind of where it really is.

Co-Authored-By: Roberto Ierusalimschy <roberto@inf.puc-rio.br>