Revert "Reduce LUAI_MAXCCALLS to avoid stack overflows"
The sieve.lua at https://www.lua.org/extras/ throws "C stack overflow" on its default setting of 500 with our C stack limit.

Raising it back to Lua's limit of 200 does not seem to have any adverse effects. I'm also not sure why exactly we lowered it in the first place, tho we should probably aim to deal with that in a better way (not taking up so much stack space) if we do encounter it again.